+++ CHANGELOG.md
@@ -1,3 +1,57 @@
+# [1.1.0-beta.6](https://github.com/touk/nk-windows/compare/v1.1.0-beta.5...v1.1.0-beta.6) (2021-08-31)
+
+
+### Bug Fixes
+
+* individual button classname ([c277006](https://github.com/touk/nk-windows/commit/c2770066ad33c2a7c7ac267650386aedcf3e42cb))
+
+# [1.1.0-beta.5](https://github.com/touk/nk-windows/compare/v1.1.0-beta.4...v1.1.0-beta.5) (2021-08-31)
+
+
+### Bug Fixes
+
+* safe (from global styles) classname ([410ff32](https://github.com/touk/nk-windows/commit/410ff325ea834db379522d8af40ea1ec653b8506))
+
+# [1.1.0-beta.4](https://github.com/touk/nk-windows/compare/v1.1.0-beta.3...v1.1.0-beta.4) (2021-08-31)
+
+
+### Features
+
+* better default footer button support for promise fail ([6ca213e](https://github.com/touk/nk-windows/commit/6ca213e0ee17caa4856ebd8ed02a3ae173006d6e))
+
+# [1.1.0-beta.3](https://github.com/touk/nk-windows/compare/v1.1.0-beta.2...v1.1.0-beta.3) (2021-08-24)
+
+
+### Bug Fixes
+
+* header buttons names instead of classnames ([c4d2a02](https://github.com/touk/nk-windows/commit/c4d2a02df304070d6e31f4b1cb230635b9cb1742))
+
+# [1.1.0-beta.2](https://github.com/touk/nk-windows/compare/v1.1.0-beta.1...v1.1.0-beta.2) (2021-08-24)
+
+
+### Bug Fixes
+
+* header button size ([7a60cc1](https://github.com/touk/nk-windows/commit/7a60cc14ee03bf963ceebb6d9956903af15258f6))
+
+
+### Features
+
+* expose header buttons styles setter ([84b7dc9](https://github.com/touk/nk-windows/commit/84b7dc904aa9dd942366961dfbc8781ec14a5585))
+
+# [1.1.0-beta.1](https://github.com/touk/nk-windows/compare/v1.0.1-beta.1...v1.1.0-beta.1) (2021-08-23)
+
+
+### Features
+
+* prevent bigger than screen windows ([4c1c307](https://github.com/touk/nk-windows/commit/4c1c3073e031028fb8b426c5eaf2b4f73d08c17e))
+
+## [1.0.1-beta.1](https://github.com/touk/nk-windows/compare/v1.0.0...v1.0.1-beta.1) (2021-08-19)
+
+
+### Bug Fixes
+
+* removed broken urlParam ([aa34eb8](https://github.com/touk/nk-windows/commit/aa34eb8853b669211693537e2437a2ca76a2a411))
+
 # 1.0.0 (2021-08-18)
 
 
@@ -17,6 +71,18 @@
 
 * some renames
 * removed LaddaButton
+
+# [1.0.0-beta.4](https://github.com/touk/nk-windows/compare/v1.0.0-beta.3...v1.0.0-beta.4) (2021-08-18)
+
+
+### Code Refactoring
+
+* rearrange code, cleanup ([102778e](https://github.com/touk/nk-windows/commit/102778e56577cf44d7b558a97877a8df056ff38d))
+
+
+### BREAKING CHANGES
+
+* some renames
 
 # [1.0.0-beta.3](https://github.com/touk/nk-windows/compare/v1.0.0-beta.2...v1.0.0-beta.3) (2021-08-18)